+++ components-starters/components-starter-metrics/src/main/java/com/zoe/framework/components/starter/metrics/configuration/MetricsProperties.java
@@ -21,10 +21,6 @@ import org.springframework.core.annotation.Order;
 public class MetricsProperties extends SerializableBean {
     private static final long serialVersionUID = 8497996001371661542L;
 
-    private Boolean enabled;
-
-    public MetricsProperties() {
-        this.enabled = true;
-    }
+    private boolean enabled = true;
 
 }