+++ components-starters/components-starter-redis/src/main/java/com/zoe/framework/components/starter/redis/configuration/redisson/GlobalRedissonProperties.java
@@ -23,7 +23,7 @@ public class GlobalRedissonProperties {
     /**
      * 是否开启redisson.
      */
-    private Boolean enabled = false;
+    private boolean enabled = false;
     /**
      * 指定redisson配置文件.
      */